re: ロジックをフックに分割
