+++ src/hooks/useConnect4Room.ts
@@ -1,0 +1,79 @@
+"use client"
+
+import { useEffect, useRef, useState } from "react";
+import { createSocket } from "@/libs/socket/client";
+import { MatchState, TurnState, handleJoinedRoomProps } from "@/types/connect4";
+import type { Socket } from "socket.io-client";
+
+export default function useConnect4Room(roomId: string) {
+	const [members, setMembers] = useState<number>(0);
+	const [playerRole, setPlayerRole] = useState<TurnState | null>(null);
+	const [matchState, setMatchState] = useState<MatchState>("waiting");
+
+	const socketRef = useRef<Socket | null>(null);
+	const membersRef = useRef<number>(0);
+	const matchStateRef = useRef<MatchState>("waiting");
+
+	useEffect(() => {
+		let pairedTimer: ReturnType<typeof setTimeout> | null = null;
+		const socket = createSocket();
+		socketRef.current = socket;
+
+		socket.connect();
+		socket.emit("joinRoom", roomId);
+
+		const handleJoinedRoom = ({ members, role }: handleJoinedRoomProps) => {
+			setMembers(members);
+			membersRef.current = members;
+			// 最初に受け取ったロールのみ採用（後から上書きしない）
+			setPlayerRole((prev) => (prev ?? role));
+		};
+
+		const handleRoomPaired = ({ roomId: pairedRoomId }: { roomId: string }) => {
+			if (pairedRoomId === roomId && matchStateRef.current === "waiting") {
+				setMatchState("matched");
+				pairedTimer = setTimeout(() => {
+					setMatchState("playing");
+				}, 2000);
+			}
+		};
+
+		const handleMembersUpdate = ({ members }: { members: number }) => {
+			setMembers(members);
+			membersRef.current = members;
+		};
+
+		socket.on("joinedRoom", handleJoinedRoom);
+		socket.on("roomPaired", handleRoomPaired);
+		socket.on("membersUpdate", handleMembersUpdate);
+
+		return () => {
+			if (pairedTimer !== null) {
+				clearTimeout(pairedTimer);
+			}
+			socket.off("joinedRoom", handleJoinedRoom);
+			socket.off("roomPaired", handleRoomPaired);
+			socket.off("membersUpdate", handleMembersUpdate);
+			socket.disconnect();
+			socketRef.current = null;
+		};
+	}, [roomId]);
+
+	useEffect(() => {
+		matchStateRef.current = matchState;
+	}, [matchState]);
+
+	const emitRestart = () => {
+		socketRef.current?.emit("restart", roomId);
+	};
+
+	return {
+		socketRef,
+		members,
+		playerRole,
+		matchState,
+		setMatchState,
+		membersRef,
+		emitRestart,
+	};
+}
