+++ src/types/connect4/index.ts
@@ -1,3 +1,5 @@
+import type { Socket } from "socket.io-client";
+
 // セルの状態を表す型（'r' = 赤、'y' = 黄、null = 空）
 export type CellState = 'r' | 'y' | null;
 
@@ -59,4 +61,25 @@ export interface OnRestartProps {
 export interface ShowTurnProps {
 	currentTurn: TurnState;
 	playerRole: TurnState | null;
+}
+
+export interface UseConnect4GameProps {
+	socketRef: React.MutableRefObject<Socket | null>;
+	matchState: MatchState;
+	playerRole: TurnState | null;
+	firstTurn: FirstState;
+	roomId: string;
+	membersRef: React.MutableRefObject<number>;
+	setMatchState: React.Dispatch<React.SetStateAction<MatchState>>;
+}
+
+export interface handleBoardUpdatedProps {
+	board: BoardState;
+	currentTurn: TurnState;
+	lastPosition: lastPositionState;
+}
+
+export interface handleJoinedRoomProps {
+	members: number;
+	role: TurnState | null;
 }
